// src/3_widgets/header/model/enums.ts
export enum LINKS {
  linkedIn = "https://www.linkedin.com/in/semyon-bizin-7a5069240?utm_source=share&utm_campaign=share_via&utm_content=profile&utm_medium=ios_app",
  gitHub = "https://github.com/Heven1y",
  telegram = "https://t.me/Heven1y",
  mail = "mailto:komrad.bizin@yandex.ru",
}

export enum TABS {
  about = "about",
  projects = "projects",
}